requirements.txt
boto==2.49.0
boto3==1.12.46
botocore==1.15.46
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
gensim==3.8.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.5
MarkupSafe==1.1.1
numpy==1.18.3
pandas==1.0.3
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==1.11.1
urllib3==1.25.9
Werkzeug==1.0.1
gunicorn
python-dotenv
